requirements.txt
psycopg2-binary==2.9.9
pandas>=1.5.0
python-dotenv==1.0.0
streamlit==1.29.0
plotly==5.17.0
altair==5.2.0
seaborn==0.13.0
matplotlib==3.8.2
openai==1.6.1
mysql-connector-python>=8.0.0
python-dateutil>=2.8.0
Flask==2.3.3
mysql-connector-python==8.1.0
Werkzeug==2.3.7
Jinja2==3.1.2
MarkupSafe==2.1.3
itsdangerous==2.1.2
click==8.1.7
blinker==1.6.3 